restructure part two
