+++ src/components/OtherStories/Login/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Page from '../../Page';
+
+const Login = () => {
+    return (
+        <Page background={'lightblue'}>
+            Login
+        </Page>
+    )
+}
+
+export default Login;
