+++ src/styles/styles.scss
@@ -1,18 +1,18 @@
 // styles.scss
 @import './base.scss';
-@import '../components/Login/style.scss';
-@import '../components/Me/style.scss';
-@import '../components/Trendex/style.scss';
-@import '../components//Era/style.scss';
-@import '../components/GenreChart/style.scss';
-@import '../components/Histogram/style.scss';
-@import '../components/SpiderGraph/style.scss';
-@import '../components/Sidebar/style.scss';
-@import '../components/Dropdown/style.scss';
-@import '../components/Dashboard/style.scss';
-@import '../components/Compare/style.scss';
-@import '../components/Survey/style.scss';
-@import '../components/NotFoundPage/style.scss';
+// @import '../components/Login/style.scss';
+// @import '../components/Me/style.scss';
+// @import '../components/Trendex/style.scss';
+// @import '../components//Era/style.scss';
+// @import '../components/GenreChart/style.scss';
+// @import '../components/Histogram/style.scss';
+// @import '../components/SpiderGraph/style.scss';
+// @import '../components/Sidebar/style.scss';
+// @import '../components/Dropdown/style.scss';
+// @import '../components/Dashboard/style.scss';
+// @import '../components/Compare/style.scss';
+// @import '../components/Survey/style.scss';
+// @import '../components/NotFoundPage/style.scss';
 @import '../components//Story/style.scss';
 @import '../components/DataStories/Decades/style.scss';
 @import '../components/DataStories/HeatMap/style.scss';
@@ -20,3 +20,4 @@
 @import '../components/DataStories/TopArtists/style.scss';
 @import '../components/DataStories/TopGenres/style.scss';
 @import '../components/DataStories/Trendex/style.scss';
+@import '../components/OtherStories/Login/style.scss';
